US59_TASK73: Understanding REST API and structure it to use later
Through this, we can understand how to call REST API and get user
specific data.
The main task to get List of Collaborators is yet to be done. But this
commit will help to understand the working of REST API better and will
be available to other team mates, so that they follow this current
structure