US59_TASK73: REST API give list of GitHub Repository Collaborators
This API will help in getting a list of Collaborators of a particular
GitHub Repository.
When Figit User is in dashboard and selects one of the GitHub projects
from a list of projects given on left side, GitHub Repository Name and
the Collaborators of that GitHub Repository appears.
This will help the Grader in knowing who all have contributed in the
particular GitHub Project.