US59-TASK73: REST API to get list of Collaborators in a GitHub Repository, searched by unique ID and Repository_Link
This commit is necssary to make sure that when user selects one
particular GitHub Repository from a list of Repositories, the Dasboard
displays the name of Project and the collaborators of that Project.
Since the Project Name may be same for more than one Repositories and
the Project Name can also be changed. So the search query should only
include id and not Project Name. Hence to get the list of collaborators
of a particular project, search query includes ID and Repository_Link as
they are unique.