US59-TASK74: Change in Database Document
As discussed, I have changed the column names of the tables
'User_Repositories' and 'Repository_Collaborators' in the document which
specifies Tables in the Database.

This change in column names will help
the UI team to display the name of github_repositories and the
information about collaborators in that repository.

We will refer this document to make sure the tables we create in
Database and the JSON data format we are using to pass data between web
pages and API are in sync.